requirements.txt
Django~=4.2.7
psycopg2~=2.9.9
environs~=9.5.0
djangorestframework~=3.14.0
